ref: 타입 분리
