+++ features/notifications/api/notification.ts
@@ -1,22 +1,8 @@
 import { apiClient } from "@/shared/utils/http";
-
-export interface TeamNotificationSettings {
-  id: number;
-  teamId: number;
-  teamName: string;
-  userId: number;
-  enableTeamAlarm: boolean;
-  enableTeamMemberNotification: boolean;
-  enableNoticeNotification: boolean;
-  enableScheduleChangeNotification: boolean;
-  enableSchedulePreNotification: boolean;
-  schedulePreNotificationMinutes: number[];
-  enableTodoChangeNotification: boolean;
-  enableTodoDeadlineNotification: boolean;
-  todoDeadlineNotificationMinutes: number[];
-  createdAt: string;
-  updatedAt: string;
-}
+import {
+  TeamNotificationSettings,
+  TeamNotificationUpdateRequest,
+} from "../types/notification.model";
 
 export async function getTeamNotificationSettings(
   teamId: number
@@ -25,18 +11,6 @@ export async function getTeamNotificationSettings(
     `notifications/teams/${teamId}`
   );
   return data;
-}
-
-export interface TeamNotificationUpdateRequest {
-  enableTeamAlarm: boolean;
-  enableScheduleChangeNotification: boolean;
-  enableSchedulePreNotification: boolean;
-  schedulePreNotificationMinutes: number[];
-  enableTodoChangeNotification: boolean;
-  enableTodoDeadlineNotification: boolean;
-  todoDeadlineNotificationMinutes: number[];
-  enableTeamMemberNotification: boolean;
-  enableNoticeNotification: boolean;
 }
 
 export async function updateTeamNotificationSettings(
